+++ m0d3l.go
@@ -1,0 +1,5 @@
+package main
+
+type ResponseData struct {
+	Stage string `json:"stage"`
+}
