+++ ∅.go
@@ -1,0 +1,6 @@
+package main
+
+func main() {
+	// not used - make the linter happy
+	EntryPoint(nil, nil)
+}
